[ION-1235] Drop LocationInSBOM field, split SBOM structs into files
